requirements.txt
Flask==1.1.1
spoonacular==3.0
google-cloud-datastore==1.7.3
